+++ apps/decodex/src/tracker/linear.rs
@@ -13,422 +13,15 @@ use crate::{
 	},
 };
 
-const LINEAR_GRAPHQL_URL: &str = "https://api.linear.app/graphql";
-const ISSUES_WITH_LABEL_QUERY: &str = r#"
-query IssuesWithLabel($labelName: String!, $after: String) {
-  issues(filter: { labels: { name: { eq: $labelName } } }, first: 50, after: $after) {
-    nodes {
-      id
-      identifier
-      title
-      creator {
-        displayName
-        name
-        email
-      }
-      description
-      priority
-      createdAt
-      updatedAt
-      state {
-        id
-        name
-      }
-      team {
-        id
-        name
-        states(first: 50) {
-          nodes {
-            id
-            name
-          }
-        }
-        labels(first: 100) {
-          nodes {
-            id
-            name
-          }
-        }
-      }
-      labels(first: 50) {
-        nodes {
-          id
-          name
-        }
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-      }
-      inverseRelations(first: 50) {
-        nodes {
-          type
-          issue {
-            id
-            identifier
-            state {
-              id
-              name
-            }
-          }
-        }
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-      }
-    }
-    pageInfo {
-      hasNextPage
-      endCursor
-    }
-  }
-}
-"#;
-const ISSUE_BY_IDENTIFIER_QUERY: &str = r#"
-query IssueByIdentifier($issueIdentifier: String!) {
-  issue(id: $issueIdentifier) {
-    id
-    identifier
-    title
-    creator {
-      displayName
-      name
-      email
-    }
-    description
-    priority
-    createdAt
-    updatedAt
-    state {
-      id
-      name
-    }
-    team {
-      id
-      name
-      states(first: 50) {
-        nodes {
-          id
-          name
-        }
-      }
-      labels(first: 100) {
-        nodes {
-          id
-          name
-        }
-      }
-    }
-    labels(first: 50) {
-      nodes {
-        id
-        name
-      }
-      pageInfo {
-        hasNextPage
-        endCursor
-      }
-    }
-    inverseRelations(first: 50) {
-      nodes {
-        type
-        issue {
-          id
-          identifier
-          state {
-            id
-            name
-          }
-        }
-      }
-      pageInfo {
-        hasNextPage
-        endCursor
-      }
-    }
-  }
-}
-"#;
-const ISSUES_BY_IDS_QUERY: &str = r#"
-query IssuesByIds($issueIds: [ID!], $after: String) {
-  issues(filter: { id: { in: $issueIds } }, first: 50, after: $after) {
-    nodes {
-      id
-      identifier
-      title
-      creator {
-        displayName
-        name
-        email
-      }
-      description
-      priority
-      createdAt
-      updatedAt
-      state {
-        id
-        name
-      }
-      team {
-        id
-        name
-        states(first: 50) {
-          nodes {
-            id
-            name
-          }
-        }
-        labels(first: 100) {
-          nodes {
-            id
-            name
-          }
-        }
-      }
-      labels(first: 50) {
-        nodes {
-          id
-          name
-        }
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-      }
-      inverseRelations(first: 50) {
-        nodes {
-          type
-          issue {
-            id
-            identifier
-            state {
-              id
-              name
-            }
-          }
-        }
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-      }
-    }
-    pageInfo {
-      hasNextPage
-      endCursor
-    }
-  }
-}
-"#;
-const ISSUE_BLOCKERS_QUERY: &str = r#"
-query IssueBlockers($issueId: String!, $after: String) {
-  issues(filter: { id: { eq: $issueId } }, first: 1) {
-    nodes {
-      inverseRelations(first: 50, after: $after) {
-        nodes {
-          type
-          issue {
-            id
-            identifier
-            state {
-              id
-              name
-            }
-          }
-        }
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-      }
-    }
-  }
-}
-"#;
-const ISSUE_COMMENTS_QUERY: &str = r#"
-query IssueComments($issueId: String!, $after: String) {
-  issue(id: $issueId) {
-    comments(first: 100, after: $after) {
-      nodes {
-        body
-        createdAt
-      }
-      pageInfo {
-        hasNextPage
-        endCursor
-      }
-    }
-  }
-}
-"#;
-const ISSUE_UPDATE_MUTATION: &str = r#"
-mutation UpdateIssue($id: String!, $input: IssueUpdateInput!) {
-  issueUpdate(id: $id, input: $input) {
-    success
-  }
-}
-"#;
-const ISSUE_UPDATE_BRIEF_MUTATION: &str = r#"
-mutation UpdateIssueBrief($id: String!, $input: IssueUpdateInput!) {
-  issueUpdate(id: $id, input: $input) {
-    success
-    issue {
-      id
-      identifier
-      title
-      creator {
-        displayName
-        name
-        email
-      }
-      description
-      priority
-      createdAt
-      updatedAt
-      state {
-        id
-        name
-      }
-      team {
-        id
-        name
-        states(first: 50) {
-          nodes {
-            id
-            name
-          }
-        }
-        labels(first: 100) {
-          nodes {
-            id
-            name
-          }
-        }
-      }
-      labels(first: 50) {
-        nodes {
-          id
-          name
-        }
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-      }
-      inverseRelations(first: 50) {
-        nodes {
-          type
-          issue {
-            id
-            identifier
-            state {
-              id
-              name
-            }
-          }
-        }
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-      }
-    }
-  }
-}
-"#;
-const ISSUE_CREATE_MUTATION: &str = r#"
-mutation CreateIssue($input: IssueCreateInput!) {
-  issueCreate(input: $input) {
-    success
-    issue {
-      id
-      identifier
-      title
-      creator {
-        displayName
-        name
-        email
-      }
-      description
-      priority
-      createdAt
-      updatedAt
-      state {
-        id
-        name
-      }
-      team {
-        id
-        name
-        states(first: 50) {
-          nodes {
-            id
-            name
-          }
-        }
-        labels(first: 100) {
-          nodes {
-            id
-            name
-          }
-        }
-      }
-      labels(first: 50) {
-        nodes {
-          id
-          name
-        }
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-      }
-      inverseRelations(first: 50) {
-        nodes {
-          type
-          issue {
-            id
-            identifier
-            state {
-              id
-              name
-            }
-          }
-        }
-        pageInfo {
-          hasNextPage
-          endCursor
-        }
-      }
-    }
-  }
-}
-"#;
-const COMMENT_CREATE_MUTATION: &str = r#"
-mutation CreateComment($input: CommentCreateInput!) {
-  commentCreate(input: $input) {
-    success
-  }
-}
-"#;
-const ISSUE_ARCHIVE_MUTATION: &str = r#"
-mutation ArchiveIssue($id: String!, $trash: Boolean) {
-  issueArchive(id: $id, trash: $trash) {
-    success
-  }
-}
-"#;
-const TEAM_LABEL_BY_NAME_QUERY: &str = r#"
-query TeamLabelByName($teamId: ID!, $labelName: String!) {
-  issueLabels(filter: { team: { id: { eq: $teamId } }, name: { eq: $labelName } }, first: 1) {
-    nodes {
-      id
-      name
-    }
-  }
-}
-"#;
+mod queries;
+
+use queries::{
+	COMMENT_CREATE_MUTATION, ISSUE_ARCHIVE_MUTATION, ISSUE_BLOCKERS_QUERY,
+	ISSUE_BY_IDENTIFIER_QUERY, ISSUE_COMMENTS_QUERY, ISSUE_CREATE_MUTATION,
+	ISSUE_UPDATE_BRIEF_MUTATION, ISSUE_UPDATE_MUTATION, ISSUES_BY_IDS_QUERY,
+	ISSUES_WITH_LABEL_QUERY, LINEAR_GRAPHQL_URL, TEAM_LABEL_BY_NAME_QUERY,
+};
+
 const LINEAR_HTTP_TIMEOUT: Duration = Duration::from_secs(30);
 const LINEAR_HTTP_CONNECT_TIMEOUT: Duration = Duration::from_secs(10);
 
@@ -1220,129 +813,4 @@ fn linear_user_display_name(user: Option<&LinearUser>) -> Option<String> {
 		.map(str::to_owned)
 }
 
-#[cfg(test)]
-mod tests {
-	use serde_json::json;
-
-	use crate::tracker::linear::{
-		GraphqlError, IssueRelationConnection, LabelConnection, LinearIssue, LinearIssueRelation,
-		LinearLabel, LinearRelatedIssue, LinearState, LinearTeam, LinearUser, PageInfo,
-		StateConnection,
-	};
-
-	#[test]
-	fn map_issue_preserves_priority_and_created_at() {
-		let issue = LinearIssue {
-			id: String::from("issue-1"),
-			identifier: String::from("PUB-101"),
-			title: String::from("Implement ordering"),
-			creator: Some(LinearUser {
-				display_name: Some(String::from("Yvette")),
-				name: Some(String::from("yvette")),
-				email: Some(String::from("yvette@example.com")),
-			}),
-			description: Some(String::from("Body")),
-			priority: Some(2),
-			created_at: String::from("2026-03-13T04:16:17.133Z"),
-			updated_at: String::from("2026-03-14T04:16:17.133Z"),
-			state: LinearState { id: String::from("state-todo"), name: String::from("Todo") },
-			team: LinearTeam {
-				id: String::from("team-1"),
-				name: String::from("Pubfi"),
-				states: StateConnection {
-					nodes: vec![LinearState {
-						id: String::from("state-todo"),
-						name: String::from("Todo"),
-					}],
-				},
-				labels: LabelConnection {
-					nodes: vec![LinearLabel {
-						id: String::from("label-needs"),
-						name: String::from("decodex:needs-attention"),
-					}],
-					page_info: None,
-				},
-			},
-			labels: LabelConnection {
-				nodes: vec![LinearLabel {
-					id: String::from("label-manual"),
-					name: String::from("decodex:manual-only"),
-				}],
-				page_info: Some(PageInfo { has_next_page: false, end_cursor: None }),
-			},
-			inverse_relations: IssueRelationConnection {
-				nodes: vec![LinearIssueRelation {
-					relation_type: String::from("blocks"),
-					issue: LinearRelatedIssue {
-						id: String::from("issue-2"),
-						identifier: String::from("PUB-102"),
-						state: LinearState {
-							id: String::from("state-progress"),
-							name: String::from("In Progress"),
-						},
-					},
-				}],
-				page_info: PageInfo { has_next_page: false, end_cursor: None },
-			},
-		};
-		let blockers = super::map_blockers(&issue.inverse_relations.nodes);
-		let mapped = super::map_issue(issue, blockers);
-
-		assert_eq!(mapped.priority, Some(2));
-		assert_eq!(mapped.author.as_deref(), Some("Yvette"));
-		assert_eq!(mapped.created_at, "2026-03-13T04:16:17.133Z");
-		assert_eq!(mapped.updated_at, "2026-03-14T04:16:17.133Z");
-		assert_eq!(mapped.blockers.len(), 1);
-		assert_eq!(mapped.blockers[0].identifier, "PUB-102");
-		assert_eq!(mapped.blockers[0].state.name, "In Progress");
-	}
-
-	#[test]
-	fn map_blockers_filters_non_blocking_relations() {
-		let blockers = super::map_blockers(&[
-			LinearIssueRelation {
-				relation_type: String::from("blocks"),
-				issue: LinearRelatedIssue {
-					id: String::from("issue-2"),
-					identifier: String::from("PUB-102"),
-					state: LinearState {
-						id: String::from("state-progress"),
-						name: String::from("In Progress"),
-					},
-				},
-			},
-			LinearIssueRelation {
-				relation_type: String::from("related"),
-				issue: LinearRelatedIssue {
-					id: String::from("issue-3"),
-					identifier: String::from("PUB-103"),
-					state: LinearState {
-						id: String::from("state-done"),
-						name: String::from("Done"),
-					},
-				},
-			},
-		]);
-
-		assert_eq!(blockers.len(), 1);
-		assert_eq!(blockers[0].identifier, "PUB-102");
-	}
-
-	#[test]
-	fn rate_limited_error_message_uses_typed_linear_extensions() {
-		let errors = vec![GraphqlError {
-			message: String::from("Too many requests"),
-			extensions: Some(json!({
-				"code": "RATELIMITED",
-				"userPresentableMessage": "API rate limit exceeded",
-				"reset": 1_777_392_000
-			})),
-		}];
-		let message =
-			super::rate_limited_error_message(&errors).expect("rate limit should classify");
-
-		assert!(message.contains("rate limited"));
-		assert!(message.contains("1777392000"));
-		assert!(message.contains("API rate limit exceeded"));
-	}
-}
+#[cfg(test)] mod tests;
